[PANA-7890] Share layer snapshot test mocks
